pyproject.toml
# pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "transformsai-ai-core"
version = "1.6.0"
description = "TransformsAI Core Library - Utilities for AI vision applications"
readme = "README.md"
authors = [
    {name = "TransformsAI", email = "contact@transformsai.com"}
]
license = { file = "LICENSE" }
# Set a more compatible Python version. Only increase this if you use newer features.
requires-python = ">=3.10,<3.13"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Core dependencies with minimum compatible versions for broad usability
dependencies = [
    "opencv-python>=4.5.0",
    "numpy>=1.26.4",
    "requests>=2.25.0",
    # "paho-mqtt>=1.6.0", # Uncomment if MQTT functionality is needed
    "psutil>=5.8.0",
    "PyYAML>=6.0",
    "fsspec>=2022.1.0",
    "tqdm>=4.62.0",
    "python-dateutil>=2.8.0",
    "pytz>=2021.1",
    "loguru>=0.7.3",
    "pydantic>=2.12.5",
]

# [project.urls]
# Homepage = "https://github.com/yourusername/transformsai-ai-core"
# Repository = "https://github.com/yourusername/transformsai-ai-core"
# Issues = "https://github.com/yourusername/transformsai-ai-core/issues"

# --- Setuptools Configuration (Replaces setup.cfg and MANIFEST.in) ---
[tool.setuptools.packages.find]
# This tells setuptools to look for packages inside the 'src' directory
where = ["src"]

[tool.setuptools]
include-package-data = true
